+++ actix-session/src/lib.rs
@@ -107,29 +107,29 @@ attached to your sessions. You can enable:
   ```
 
 - a Redis-based backend via [`redis-rs`](https://docs.rs/redis-rs), [`RedisSessionStore`], using
-  the `redis-rs-session` feature flag.
+  the `redis-session` feature flag.
 
   ```toml
   [dependencies]
   # ...
-  actix-session = { version = "...", features = ["redis-rs-session"] }
+  actix-session = { version = "...", features = ["redis-session"] }
   ```
 
-  Add the `redis-rs-tls-session` feature flag if you want to connect to Redis using a secured
+  Add the `redis-session-native-tls` feature flag if you want to connect to Redis using a secured
   connection (via the `native-tls` crate):
 
   ```toml
   [dependencies]
   # ...
-  actix-session = { version = "...", features = ["redis-rs-tls-session"] }
+  actix-session = { version = "...", features = ["redis-session-native-tls"] }
   ```
 
-  If you instead prefer depending on `rustls`, use the `redis-rs-tls-session-rustls` feature flag:
+  If you instead prefer depending on `rustls`, use the `redis-session-rustls` feature flag:
 
   ```toml
   [dependencies]
   # ...
-  actix-session = { version = "...", features = ["redis-rs-tls-session-rustls"] }
+  actix-session = { version = "...", features = ["redis-session-rustls"] }
   ```
 
 You can implement your own session storage backend using the [`SessionStore`] trait.